+++ R/misc.r
@@ -176,6 +176,26 @@ add_lists <- function(l1, l2) {
   return(result)
 }
 
+#' Determine duplicate elements including their first occurrence.
+#' @description If an element is duplicated, all of its occurrence will be labeled `TRUE`.
+#'   Useful to list and compare all duplicates.
+#' @param x A vector.
+#'
+#' @returns A logical value.
+#' @export
+#' @examples
+#' indicate_duplicates(c(1, 2, NA, NA, 1))
+#' indicate_duplicates(c(1, 2, 3, 4, 4))
+#' 
+#' # Useful to check duplicates in data frames.
+#' df <- data.frame(
+#'   id = c(1, 2, 1, 2, 3), year = c(2010, 2011, 2010, 2010, 2011),
+#'   value = c(1, 2, 3, 4, 5)
+#' )
+#' df[indicate_duplicates(df[, c("id", "year")]), ]
+indicate_duplicates <- function(x) {
+  c(duplicated(x) | duplicated(x, fromLast = TRUE))
+}
 
 #' Replacing elements in a vector
 #' @param x A vector.